Move root params regression to dedicated e2e
