+++ test/e2e/app-dir/segment-cache/vary-params/vary-params.test.ts
@@ -669,16 +669,9 @@ describe('segment cache - vary params', () => {
     // Root params accessed via rootParams() are tracked in varyParams.
     // Different param values require separate prefetches.
     let act: ReturnType<typeof createRouterAct>
-    const segmentPrefetchBodies: Array<Promise<string>> = []
     const browser = await next.browser('/root-params', {
       beforePageLoad(p: Playwright.Page) {
         act = createRouterAct(p)
-        p.on('response', (response) => {
-          const request = response.request()
-          if (request.headers()['next-router-segment-prefetch']) {
-            segmentPrefetchBodies.push(response.text().catch(() => ''))
-          }
-        })
       },
     })
 
@@ -703,16 +696,5 @@ describe('segment cache - vary params', () => {
       },
       { includes: 'Root param page content - param: bbb' }
     )
-
-    const settledSegmentPrefetchBodies = await Promise.all(
-      segmentPrefetchBodies
-    )
-
-    expect(settledSegmentPrefetchBodies.length).toBeGreaterThan(0)
-    expect(
-      settledSegmentPrefetchBodies.some((body) =>
-        body.includes('%5BrootParam%5D')
-      )
-    ).toBe(false)
   })
 })
